更新Findx.WebHost组件使用示例

+++ src/Findx.WebHost/Messaging/CancelOrderCommand.cs
@@ -8,11 +8,11 @@ namespace Findx.WebHost.Messaging
         {
         }
 
-        public CancelOrderCommand(int orderId)
+        public CancelOrderCommand(long orderId)
         {
             OrderId = orderId;
         }
 
-        public int OrderId { get; private set; }
+        public long OrderId { get; private set; }
     }
 }
